Addressing potential XSS exploit #1
In previous verison, JS values are treated as-is and just placed in HTML content of HTML elements. This might lead to XSS exploit as described in #1. To address this, text is used instead of html for plain text fields. New option RAW_HTML is added to switch to previous behaviour.